QKeySequence: replace an inefficient QList with QVector
QModifKeyName is larger than a pointer, so holding it
in a QList is horribly inefficient.

Fix by marking it movable and using QVector instead.

The logic for filling the lists-turned-vectors is a bit
fishy in that it is not thread-safe. Maybe it doesn't
have to, it's not marked as \reentrant.

Change-Id: I8421e6d8b980eff022badfe3c61b3537783b5cfa
Reviewed-by: Olivier Goffart (Woboq GmbH) <ogoffart@woboq.com>