Remove QPlatformScreenPageFlipper.
Use of this was removed from QtWayland quite a long time ago, which was the only
public user of this API.

Furthermore, it isn't easily possible to implement any use of this API without
full control of the graphics stack (a very rare occurrence) and there is no
public demonstration of this.

There is ongoing research to provide a better replacement for this in the form
of QPlatformHardwareCompositor.

Change-Id: I80d666a5b465aa80f73fed6c44838ce7210bbd30
Reviewed-by: Andrew Knight <andrew.knight@digia.com>
Reviewed-by: Gunnar Sletta <gunnar.sletta@jollamobile.com>